fix(wifi): %zu format bug + minimal flash-credential boot wiring (038-006 slice)
(1) WireHandler::execWifiCred()'s bare enumeration used %zu on a size_t
index. This target's embedded newlib-nano printf does not support the
z length modifier -- MEASURED gopiv 2026-09-09,
captures/wifi-credential-store-20260909/: `WIFICRED SET 0 TestNet038
secretpw038 #2` wrote fine, but `WIFICRED #3` enumerated
`wificred zu 0` instead of `wificred 0 TestNet038 1`. Host tests
passed because the host's own printf does support %zu. Fixed with an
explicit static_cast<unsigned> + %u, matching replyErr()'s existing
precedent. Added
tests/host/test_no_percent_z_format_specifier_source_pin.py, a
repo-wide grep-based guard forbidding %z/%zu/%zd/etc. anywhere in
src/, so this class of bug cannot pass host tests silently again.

(2) Minimal boot wiring so a stored credential actually gets used --
a deliberately REDUCED slice of ticket 006. Protocol::serviceWifi()'s
lazy-begin now prefers the FIRST occupied WifiCredentialStore slot
over setupWifi()'s wifiCredsExplicit_ credential, which in turn beats
the baked kWifiSsid/kWifiPassword. An empty store leaves behavior
byte-for-byte unchanged. emitWifiDebug()'s credsrc= field gains value
2 for the flash case (0=baked, 1=setupWifi(), 2=flash), documented in
both protocol.cpp's own comment and docs/robot-connections.md.
WIFICRED SET writes flash immediately but only takes effect at the
next boot -- commented at wire_adapter.cpp's wifiCredSet(), whose
ack/err reply already made no claim otherwise.

Deliberately OUT of this slice, per explicit stakeholder
deprioritization: ticket 005's WifiJoinSequencer and any
multi-credential list-walking/retry-on-failure behavior. Only the
FIRST occupied slot is ever tried; sprint.md's Architecture section
(and its 2026-09-10 Revision) is unchanged and still describes the
full design this deviates from for this interim release.

Extended tests/host/test_setupwifi_precedence_source_pin.py with two
new source-pin tests covering the flash-first precedence and the
credsrc=2 mapping (protocol.cpp cannot be host-compiled -- pxt.h --
so these are source-shape pins, same limitation as the rest of that
file).

Scoped host suite run (466 tests): test_wire_grammar.py,
test_wifi_credential_store.py, test_wifi_join_error_debug_source_pin.py,
test_no_units_in_identifiers_source_pin.py,
test_setupwifi_precedence_source_pin.py,
test_no_percent_z_format_specifier_source_pin.py,
test_wifi_flash_port_write_sense_source_pin.py, test_wifi_link.py,
test_config_surface_single_source.py, test_config_descriptor_table.py,
test_wire_motion_verbs.py -- all passing.

+++ src/comms/wire_handler.cpp
@@ -1098,7 +1098,19 @@ void WireHandler::execWifiCred(char** fields, size_t fieldCount, uint32_t id,
       // -- flash content is not trusted to be free of '\n'/'\r' any
       // more than a RunRegistry entry is.
       sanitizeLineText(ssid, sanitizedSsid, sizeof(sanitizedSsid));
-      snprintf(buf, sizeof(buf), "wificred %zu %s %d\n", i, sanitizedSsid,
+      // %zu is NOT supported by this target's printf -- MEASURED gopiv
+      // 2026-09-09, captures/wifi-credential-store-20260909/: the
+      // embedded newlib-nano printf emits the two characters "zu"
+      // literally and shifts every remaining argument, so
+      // `WIFICRED #3` enumerated `wificred zu 0` instead of
+      // `wificred 0 TestNet038 1`. Host tests pass regardless because
+      // the host's own printf DOES support %zu, which is exactly why
+      // this needs a source-pin guard
+      // (test_no_percent_z_format_specifier_source_pin.py) and not just
+      // this one fix. Explicit cast to unsigned + %u, matching
+      // replyErr()'s own `static_cast<unsigned>(code)` precedent above.
+      snprintf(buf, sizeof(buf), "wificred %u %s %d\n",
+               static_cast<unsigned>(i), sanitizedSsid,
                hasPassword ? 1 : 0);
       writeLine(buf);
     }
